VT2-32-회원가입 Srping AI 서비스 테스트 코드 작성
